Take another step towards pushing float data types as the default for all operaitons possible.  This could significantly improve performance for platforms that accelerate floating point operations, but not double operations.

+++ src/nav_openloop/glocal.hxx
@@ -4,4 +4,4 @@
 // at the wgs84 reference elliposid using the Somigliana model and
 // makes corrections for altitude. 
 
-Vector3d local_gravity( double lat_rad, double alt_m );
+Vector3f local_gravity( double lat_rad, float alt_m );
